arch(#1948): migrate scoring.py to src/ (slice 5)
Fifth concrete migration slice for #1948. Mirrors PR #2018 (phone_utils,
slice 1), PR #2020 (mini_app observability, slice 2), PR #2024
(content_loader, slice 3), and PR #2030 (kommo CRM stack, slice 4):
canonical home moves to src/, telegram_bot/ keeps a thin re-export shim,
and src/api/main.py's lazy import switches to the canonical path.

Files moved (via git mv to preserve history)
--------------------------------------------
- telegram_bot/scoring.py to src/scoring.py
  Pure stdlib helpers (numbers, typing) that write Langfuse scores from
  RAG result dicts. No telegram_bot.* references at module scope. 402
  lines. Public API: compute_checkpointer_overhead_proxy_ms, score,
  write_crm_scores, write_history_scores, write_langfuse_scores.

Files added (re-export shim)
----------------------------
- telegram_bot/scoring.py (33 lines) re-exports every public callable
  from src.scoring with object identity. Existing internal callers
  (telegram_bot/agents/rag_tool.py, telegram_bot/handlers/command_handlers.py,
  telegram_bot/pipelines/client.py, telegram_bot/bot.py,
  scripts/validate_traces.py) keep working unchanged because every
  public name still resolves at the original path with the same
  identity.

src/api
-------
src/api/main.py changes the lazy import inside _run_rag_pipeline from
'from telegram_bot.scoring import write_langfuse_scores' to
'from src.scoring import write_langfuse_scores'. Drops the last
forbidden import in src/api/main.py for scoring per #1948.

Test patches
------------
4 test patches updated from telegram_bot.scoring.write_langfuse_scores
to src.scoring.write_langfuse_scores. After src/api/main.py imports
from the canonical, the patch site must follow:
- tests/unit/api/test_rag_api_runtime.py (3 sites at lines 113, 357, 426)
- tests/unit/test_validate_aggregates.py (line 1108)

Drift guard
-----------
tests/contract/test_issue_1948_scoring_slice_contract.py asserts:

  1. src/scoring.py exists and is non-empty.
  2. Canonical does NOT import from telegram_bot.* at module level.
  3. Canonical exposes the full public API (5 callables).
  4. The telegram_bot shim re-exports each public callable with
     object-identity (is-equal) - same guarantee as PR #2018, PR #2024,
     PR #2030.
  5. src/api/main.py imports from src.scoring and contains zero
     references to telegram_bot.scoring.

Validation
----------
  uv run pytest \
    tests/contract/test_issue_1948_scoring_slice_contract.py \
    tests/unit/agents/test_nurturing_observability.py \
    tests/unit/graph/test_call_limits.py \
    tests/unit/observability/ \
    tests/unit/test_validate_aggregates.py \
    -q --no-cov                                          to 273 passed
  ruff check / ruff format --check on touched files       to clean

What does NOT change
--------------------
- No production behaviour change beyond import paths; the shim
  preserves identity (is-equal) of every public callable.
- Score names, ranges, conditional emission rules, and the
  _QUERY_TYPE_SCORE mapping all keep their original contract verbatim.

Refs #1948 (does not close - src/api/main.py still imports
telegram_bot.graph.{config,graph,state}, telegram_bot.integrations.cache,
and telegram_bot.services.qdrant; those need separate slices since they
overlap with #1538 voice-path migration to create_agent SDK).
